Request/Response logging
Made good use of .net 6 builtin HttpLogging, also added Serilog for
structured logging, it is simply to good to pass upp. I did make an
attempt to achieve filtering via the builtin logging framework but short
of writing my own provider there seems to way. In stark contrast to this
serilog provides an outofthebox way to filter, in this case i wanted to
filter out all calls to the swagger endpoint.